+++ test/snek/ruleset_test.exs
@@ -1,0 +1,5 @@
+defmodule RulesetTest do
+  use ExUnit.Case
+  alias Snek.Ruleset
+  doctest Ruleset, import: true
+end